+++ packages/generator-odata-downloader/src/data-download/odata-download-generator.ts
@@ -14,6 +14,7 @@ import { writeExternalServiceMetadata } from '@sap-ux/odata-service-writer';
 import { DirName, getMockServerConfig } from '@sap-ux/project-access';
 import type { IVSCodeExtLogger, LogLevel } from '@vscode-logging/logger';
 import { join } from 'node:path';
+// @ts-expect-error: No types available, breaks no implicit any rule
 import prettifyXml from 'prettify-xml';
 import type { GeneratorOptions } from 'yeoman-generator';
 import Generator from 'yeoman-generator';