+++ packages/system-access/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @sap-ux/system-access
 
+## 0.6.66
+
+### Patch Changes
+
+- Updated dependencies [a854433]
+    - @sap-ux/axios-extension@1.25.24
+
+## 0.6.65
+
+### Patch Changes
+
+- 5d452e5: fix(deps): update dependency nock to v14
+- Updated dependencies [5d452e5]
+- Updated dependencies [55417bb]
+    - @sap-ux/axios-extension@1.25.23
+    - @sap-ux/btp-utils@1.1.10
+    - @sap-ux/store@1.5.10
+
 ## 0.6.64
 
 ### Patch Changes